order function

+++ Models/Product.cs
@@ -7,8 +7,8 @@ namespace WebServiceShopping.Models
         public int idsp { get; set; }
         public int idloaisp { get; set; }
         public string tensp { get; set; }
-        public int gianhap { get; set; }
-        public int giaban { get; set; }
+        public Decimal gianhap { get; set; }
+        public Decimal giaban { get; set; }
         public string thongtinsp { get; set; }
         public int slsanpham { get; set; }
         public DateTime ngaynhaphang { get; set; }
